\name{print.SurvAM}
\alias{print.SurvAM}
\title{Print an object of class SurvAM--output from survEstMeasures}
\usage{
  print.SurvAM(x)
}
\arguments{
  \item{x,output}{from the function survEstMeasures}
}
\value{
  NULL
}
\description{
  Print an object of class SurvAM--output from
  survEstMeasures
}

