\name{print.SurvAM}
\alias{print.SurvAM}
\title{Print an object of class SurvAM--output from survEstMeasures}

\description{
  S3 method for class ''SurvAM''
}
\usage{
#S3 method for class ''trtsel''
\method{print}{SurvAM}(x, ...)
}

\arguments{
  \item{x}{output from the function survEstMeasures}
  \item{\dots}{NA}
}
\value{
  NULL
}

