\name{print.SurvAM}
\alias{print.SurvAM}
\title{Print an object of class SurvAM--output from survEstMeasures}
\usage{
  print(x, ...)
}
\arguments{
  \item{x}{output from the function survEstMeasures}
  \item{\dots}{NA}
}
\value{
  NULL
}
\description{
  Print an object of class SurvAM--output from
  survEstMeasures
}

